+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/hyperlink button/PrivateDisabledHyperlinkButton.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import HyperlinkButton from 'terra-hyperlink/lib/HyperlinkButton';
+
+export default () => (
+  <div role="main">
+    <p>
+      <HyperlinkButton isDisabled onClick={() => {}} text="Hyperlink Button" />
+    </p>
+  </div>
+);